requirements.txt
google-generativeai>=0.3.0
imap-tools>=1.5.0
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0
